Addition of amavisd-new, clamav and spamassassin
Added resource notes and improved comments for clarity throughout the
script.

Improved user and directory checks pre-run.

Improved installTemplate function to enable usage on templates that do
not contain tags.  Updated all prior usage of installTemplate and added
to new locations that can now use it.

Modified the ps3 for prompt clarity.

Added creation of a postmaster@[domain] email account which will recieve
notifications for the mailserver.

Correction to postfix main.cf where the domain was not being set
properly for one of the variables.

Added installation of rpmforge for amavisd-new and clamav (spamassassin
from rpmforge was having issues with dependencies).  Script now installs
amavisd-new, clamav and spamassassin then integrates with postfix.

Initialized a virtualhost in apache for the domain that is set during
installation to fix errors during apache startup.

Lowered iptables logging level.