Set mail.domain.com as default for mail server
Instead of using host.domain.com, added mail.domain.com to hosts file
and mail server systems.  The idea is to make it easier to have multiple
mail servers across a domain and not tie the setup to a particular
host.

Changed SSH port default from 'ssh' to '22' as it is used to set the
sshd config file, sshd does not recognize 'ssh' and will lock you out!

Added insertion into domains mysql table for the default domain,
postmaster account will not work without it.