Functional virtualhosts with ssl, bug fixes
Cleaned the crap out of the httpd.conf template.  Removed a lot of the
notes to hopefully make it easier to find and read directives.  Moved as
many directives as appropriate out of the /var/www/html directory to
apply server wide.  Created a /var/www/vhosts directory for all
virtualhosts.  Enabled persistent connections.

Added ssl.conf template as a bit of it was moved over to httpd.conf.  In
particular the first ssl virtualhost declaration.

Corrections to the vhost.conf template, apache was not starting up due
to issues with this file.

Addition of placeholder index.php files into /var/www/html and
/var/www/vhosts/domain.com.

Cleaned up the task list.

Got the distro check to actually work and check for the appropriate
version as well.